% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_epsg}
\alias{hs_epsg}
\title{Get the EPSG code of a hyperspectral image}
\usage{
hs_epsg(filename)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}
}
\value{
character representation of an epsg code, e.g., '4326'
}
\description{
This function returns the \href{http://www.epsg.org/}{EPSG} code that defines
the spatial coordinate reference system of an image.
}
