% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_extent}
\alias{hs_extent}
\title{Get the spatial extent of a hyperspectral image}
\usage{
hs_extent(filename)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}
}
\value{
a raster::extent object that contains the min and max x and y
coordinates of a hyperspectral image
}
\description{
Get the spatial extent of a hyperspectral image
}
\examples{
path_to_file <- system.file('extdata', 'ex.h5', package = 'neonaop')
hs_extent(path_to_file)

}
