% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_extract_pts}
\alias{hs_extract_pts}
\title{Efficiently extract spectra using a SpatialPointsDataFrame}
\usage{
hs_extract_pts(filename, pts, bands)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}

\item{pts}{a \code{SpatialPointsDataFrame} object defining points where data will
be extracted}

\item{bands}{Indices of bands to read (integer)}
}
\description{
The \code{hs_extract_pts} function efficiently extracts hyperspectral reflectance
data at spatial points. This is more efficient than reading the entire
raster using \code{hs_read}, then using \code{raster::extract}.
}
\details{
Points that are not contained within the extent of the raster will return NA
values.
}
