% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_proj4string}
\alias{hs_proj4string}
\title{Get the proj4string for a hyperspectral image}
\usage{
hs_proj4string(filename)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}
}
\value{
character proj4string representation
}
\description{
This returns the proj4string representation of the coordinate reference
system of an image.
}
