% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_read}
\alias{hs_read}
\title{Read hyperspectral imagery from the NEON AOP}
\usage{
hs_read(filename, bands, crop = FALSE)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}

\item{bands}{Indices of bands to read (integer)}

\item{crop}{Optional extent object to use in cropping}
}
\value{
Raster* object containing hyperspectral data
}
\description{
The \code{hs_read} function reads hyperspectral imagery from NEON's Aerial
Observation Platform, and returns a Raster* object.
}
\details{
You can learn more the NEON AOP hyperspectral imagery at:
https://www.neonscience.org/data-collection/airborne-remote-sensing
}
