% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs_read.R
\name{hs_wavelength}
\alias{hs_wavelength}
\title{Get wavelengths from a hyperspectral image}
\usage{
hs_wavelength(filename, bands)
}
\arguments{
\item{filename}{Path to an .h5 file containing L3 hyperspectral data (char)}

\item{bands}{Indices of bands to read (integer)}
}
\value{
character vector containing band indices and wavelengths
}
\description{
The \code{hs_wavelength} returns a character vector containing band indices and
wavelengths rounded to the nearest nanometer, e.g., band1_384nm, band2_389nm,
where band... gives the band index (first band, second band, etc.) and
...nm gives the wavelength of that band in nanometers.
}
